Refactor hashset tests and fix minor issues
Refactored test_hashset.c for improved code style and static functions. Updated material.c to clarify descriptor set destruction. Fixed function signature in pipeline.c. Corrected method name in Lua engine teardown. Updated Xcode project structure for tkn source organization. Enabled window close termination in AppDelegate.

+++ tkn/src/pipeline.c
@@ -1,5 +1,5 @@
 #include "gfxCore.h"
-static size_t getSizeOfVkFormat(VkFormat format)
+static uint32_t getSizeOfVkFormat(VkFormat format)
 {
     switch (format)
     {
